weird flashing, changes state, doesn't persist

+++ src/main/java/com/jhuoose/monadic/models/Lesson.java
@@ -3,10 +3,10 @@ package com.jhuoose.monadic.models;
 import java.util.ArrayList;
 
 public class Lesson {
-    double ID;
+    int ID;
     ArrayList<LessonElement> lessonElements;
 
-    public Lesson(double ID, ArrayList<LessonElement> lessonElements) {
+    public Lesson(int ID, ArrayList<LessonElement> lessonElements) {
         this.ID = ID;
         this.lessonElements = lessonElements;
     }
@@ -15,7 +15,7 @@ public class Lesson {
         return ID;
     }
 
-    public void setID(double ID) {
+    public void setID(int ID) {
         this.ID = ID;
     }
